refactor to make code dry and sensible utils
