+++ akd_ext/tools/worldview/__init__.py
@@ -1,14 +1,13 @@
 from akd_ext.tools.worldview.permalink import (
+    LayerSpec,
     WorldviewPermalinkInputSchema,
     WorldviewPermalinkOutputSchema,
     WorldviewPermalinkTool,
 )
-from akd_ext.tools.worldview.utils import LayerSpec, build_worldview_permalink
 
 __all__ = [
     "LayerSpec",
     "WorldviewPermalinkInputSchema",
     "WorldviewPermalinkOutputSchema",
     "WorldviewPermalinkTool",
-    "build_worldview_permalink",
 ]
